build(ray_marching): move the ray_marching code into its own module
