+++ poom/pooma/math.pxd
