+++ poom/pooma/ray_march.pyi
